+++ weblate/urls.py
@@ -874,7 +874,7 @@ real_patterns = [
         name='hook-project',
     ),
     url(
-        r'^hooks/(?P<service>github|gitlab|bitbucket|pagure|azure|gitea)/$',
+        r'^hooks/(?P<service>github|gitlab|bitbucket|pagure|azure|gitea|gitee)/$',
         weblate.trans.views.hooks.vcs_service_hook,
         name='webhook',
     ),